Refactors Endpoints and JWPlayer creation out of bloated VM squash
