--- JWPlayer-Developer-Demo/ObjCPlayerViewController.h
@@ -1,13 +1,0 @@
-//
-//  PlayerViewController.h
-//  JWPlayerSDKDemo
-//
-//  Created by Max Mikheyenko on 1/5/15.
-//  Copyright (c) 2015 JWPlayer. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-@interface ObjCPlayerViewController : UIViewController
-
-@end
